fix: number of windows (cycles) calculated from delta
addded a info function to window object to print out first and last model times

The model start time follows the window start time of +1 seconds. Not sure if we
should do this or have the model time +0 seconds.  For tiegcm it is irrelevant
since the model input is year, day, hour, minute