localizing gw_tau with infinite horizontal radius
gw_tau’s lon, lat are set to the location of the observation. This gives you
an infinite localization radius in the horizontal by ensuring that
the horizontal distance from an observation to gw_tau is always = 0

The vertical location of gw_tau is set with the namelist item gw_tau_vert
The coordinate is whatever you are using for vertical localization (e.g. pressure, scaleheight)

fixed bug in get_state_meta_data: removed early return. qty now returned for gw_tau