add a test case for the obs_errors, and fix a bug that only shows up with debugging flags
add a test case to provoke a bug in the ncep obs_error mod.  this only shows up
if you build with the debug compiler flags, which trap on taking log(0).  the default
optimization level just returns 0 for log(0) which gives the right answer.

this fixes the confusion that if you turn on debugging to diagnose some other problem,
the code will stop here instead.